refactor(spx-gui): rename EnrichGateButton to EnrichableSubmitButton and polish implementation
